requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
deepl==1.16.1
idna==3.6
invoke==2.2.0
packaging==23.2
pillow==10.2.0
PyMuPDF==1.23.16
PyMuPDFb==1.23.9
pypdf==3.17.4
pytesseract==0.3.10
requests==2.31.0
urllib3==2.1.0
